l-01 change the type of chainId to uint256

+++ contracts/ChainlinkLightClient.sol
@@ -28,11 +28,11 @@ contract ChainlinkLightClient is ILightClient, IChainlinkLightClient, Ownable {
     address public immutable GATEWAY;
 
     // chainId => height => account => storageRoot
-    mapping(uint32 => mapping(uint256 => mapping(address => bytes32)))
+    mapping(uint256 => mapping(uint256 => mapping(address => bytes32)))
         public approvedStorageRoots;
 
     // chainId => height => stateRoot
-    mapping(uint32 => mapping(uint256 => bytes32)) public approvedStateRoots;
+    mapping(uint256 => mapping(uint256 => bytes32)) public approvedStateRoots;
 
     // wallet => isWhitelisted
     mapping(address => bool) public whitelist;
@@ -47,7 +47,7 @@ contract ChainlinkLightClient is ILightClient, IChainlinkLightClient, Ownable {
      * @param root State root
      */
     struct Proof {
-        uint32 dstChainId;
+        uint256 dstChainId;
         uint256 height;
         bytes proof;
     }
@@ -83,7 +83,7 @@ contract ChainlinkLightClient is ILightClient, IChainlinkLightClient, Ownable {
      * @param root State root
      */
     event UpdateStateRoot(
-        uint32 indexed chainId,
+        uint256 indexed chainId,
         uint256 indexed height,
         bytes32 root
     );
